Remove AnyMemoService in favor of direct notification manipulation
Also fixed IllegalStateException crash in Oreo
